Split device/gfx.rs into multiple modules
