+++ webrender/src/debug_render.rs
@@ -16,7 +16,8 @@ cfg_if! {
         use device::Program;
     } else {
         use device::{PrimitiveType, ProgramId as Program, ShaderKind};
-        use vertex_types;
+        #[path = "device/gfx/vertex_types.rs"]
+        mod vertex_types;
     }
 }
