+++ webrender/src/device/gfx/blend_state.rs
@@ -1,0 +1,133 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+use hal::pso::{BlendOp, BlendState, Factor};
+
+pub(super) const ALPHA: BlendState = BlendState::On {
+    color: BlendOp::Add {
+        src: Factor::SrcAlpha,
+        dst: Factor::OneMinusSrcAlpha,
+    },
+    alpha: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::One,
+    },
+};
+
+pub(super) const PREMULTIPLIED_DEST_OUT: BlendState = BlendState::On {
+    color: BlendOp::Add {
+        src: Factor::Zero,
+        dst: Factor::OneMinusSrcAlpha,
+    },
+    alpha: BlendOp::Add {
+        src: Factor::Zero,
+        dst: Factor::OneMinusSrcAlpha,
+    },
+};
+
+pub(super) const MAX: BlendState = BlendState::On {
+    color: BlendOp::Max,
+    alpha: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::One,
+    },
+};
+
+pub(super) const MIN: BlendState = BlendState::On {
+    color: BlendOp::Min,
+    alpha: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::One,
+    },
+};
+
+pub(super) const SUBPIXEL_PASS0: BlendState = BlendState::On {
+    color: BlendOp::Add {
+        src: Factor::Zero,
+        dst: Factor::OneMinusSrcColor,
+    },
+    alpha: BlendOp::Add {
+        src: Factor::Zero,
+        dst: Factor::OneMinusSrcAlpha,
+    },
+};
+
+pub(super) const SUBPIXEL_PASS1: BlendState = BlendState::On {
+    color: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::One,
+    },
+    alpha: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::One,
+    },
+};
+
+pub(super) const SUBPIXEL_WITH_BG_COLOR_PASS0: BlendState = BlendState::On {
+    color: BlendOp::Add {
+        src: Factor::Zero,
+        dst: Factor::OneMinusSrcColor,
+    },
+    alpha: BlendOp::Add {
+        src: Factor::Zero,
+        dst: Factor::One,
+    },
+};
+
+pub(super) const SUBPIXEL_WITH_BG_COLOR_PASS1: BlendState = BlendState::On {
+    color: BlendOp::Add {
+        src: Factor::OneMinusDstAlpha,
+        dst: Factor::One,
+    },
+    alpha: BlendOp::Add {
+        src: Factor::Zero,
+        dst: Factor::One,
+    },
+};
+
+pub(super) const SUBPIXEL_WITH_BG_COLOR_PASS2: BlendState = BlendState::On {
+    color: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::One,
+    },
+    alpha: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::OneMinusSrcAlpha,
+    },
+};
+
+// This requires blend color to be set
+pub(super) const SUBPIXEL_CONSTANT_TEXT_COLOR: BlendState = BlendState::On {
+    color: BlendOp::Add {
+        src: Factor::ConstColor,
+        dst: Factor::OneMinusSrcColor,
+    },
+    alpha: BlendOp::Add {
+        src: Factor::ConstAlpha,
+        dst: Factor::OneMinusSrcAlpha,
+    },
+};
+
+pub(super) const SUBPIXEL_DUAL_SOURCE: BlendState = BlendState::On {
+    color: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::OneMinusSrc1Color,
+    },
+    alpha: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::OneMinusSrc1Alpha,
+    },
+};
+
+pub(super) const OVERDRAW: BlendState = BlendState::On {
+    color: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::OneMinusSrcAlpha,
+    },
+    alpha: BlendOp::Add {
+        src: Factor::One,
+        dst: Factor::OneMinusSrcAlpha,
+    },
+};
+
